Initial port to 1.20.2

+++ common/src/main/java/net/creeperhost/polylib/client/modulargui/lib/GuiRender.java
@@ -1126,11 +1126,11 @@ public class GuiRender extends LegacyRender {
         }
     }
 
-    private void bufferDynamic(VertexConsumer builder, Matrix4f mat, TextureAtlasSprite tex, int x, int y, double textureX, double textureY, int width, int height, float red, float green, float blue, float alpha) {
+    private void bufferDynamic(VertexConsumer builder, Matrix4f mat, TextureAtlasSprite tex, int x, int y, float textureX, float textureY, int width, int height, float red, float green, float blue, float alpha) {
         int w = tex.contents().width();
         int h = tex.contents().height();
         //@formatter:off
-        builder.vertex(mat, x,         y + height, 0).color(red, green, blue, alpha).uv(tex.getU((textureX / w) * 16D),          tex.getV(((textureY + height) / h) * 16)).endVertex();
+        builder.vertex(mat, x,         y + height, 0).color(red, green, blue, alpha).uv(tex.getU((textureX / w) * 16F),          tex.getV(((textureY + height) / h) * 16)).endVertex();
         builder.vertex(mat, x + width, y + height, 0).color(red, green, blue, alpha).uv(tex.getU(((textureX + width) / w) * 16), tex.getV(((textureY + height) / h) * 16)).endVertex();
         builder.vertex(mat, x + width, y,          0).color(red, green, blue, alpha).uv(tex.getU(((textureX + width) / w) * 16), tex.getV(((textureY) / h) * 16)).endVertex();
         builder.vertex(mat, x,         y,          0).color(red, green, blue, alpha).uv(tex.getU((textureX / w) * 16),           tex.getV(((textureY) / h) * 16)).endVertex();
